This only adds the commit e564ea674cd09ff8c5ad17afa0afd67bf350ddd2

   Bug 7358 - GPSC : inode count over int limit, added TMPDIR and TRUE_HOST
   default values if not defined

+++ src/l2d2_commun.c
@@ -74,7 +74,7 @@ void get_time (char *btime , int mode )
  *  Author      : R.Lahlou , cmoi 2012
  *  Description : get the inode of a path , file , ...
  */ 
-int get_Inode ( const char *pathname )
+size_t get_Inode ( const char *pathname )
 {
      char *path_status=NULL;
      char resolved[MAXPATHLEN];
